// src/contracts/newport/typechain/OptionMarketViewer.ts
export declare namespace OptionMarketViewer {
// ...
  export type OptionMarketAddressesStructOutput = [
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string,
    string
  ] & {
    liquidityPool: string;
    liquidityToken: string;
    greekCache: string;
    optionMarket: string;
    optionMarketPricer: string;
    optionToken: string;
    shortCollateral: string;
    poolHedger: string;
    quoteAsset: string;
    baseAsset: string;
  };
// ...
  export type StrikeViewStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    OptionGreekCache.StrikeGreeksStructOutput,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    strikeId: BigNumber;
    boardId: BigNumber;
    strikePrice: BigNumber;
    skew: BigNumber;
    forceCloseSkew: BigNumber;
    cachedGreeks: OptionGreekCache.StrikeGreeksStructOutput;
    baseReturnedRatio: BigNumber;
    longCallOpenInterest: BigNumber;
    longPutOpenInterest: BigNumber;
    shortCallBaseOpenInterest: BigNumber;
    shortCallQuoteOpenInterest: BigNumber;
    shortPutOpenInterest: BigNumber;
  };
// ...
  export type BoardViewStructOutput = [
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    boolean,
    BigNumber,
    BigNumber,
    BigNumber,
    OptionGreekCache.NetGreeksStructOutput,
    OptionMarketViewer.StrikeViewStructOutput[]
  ] & {
    market: string;
    boardId: BigNumber;
    expiry: BigNumber;
    baseIv: BigNumber;
    priceAtExpiry: BigNumber;
    isPaused: boolean;
    varianceGwavIv: BigNumber;
    forceCloseGwavIv: BigNumber;
    longScaleFactor: BigNumber;
    netGreeks: OptionGreekCache.NetGreeksStructOutput;
    strikes: OptionMarketViewer.StrikeViewStructOutput[];
  };
// ...
  export type MarketParametersStructOutput = [
    OptionMarket.OptionMarketParametersStructOutput,
    LiquidityPool.LiquidityPoolParametersStructOutput,
    LiquidityPool.CircuitBreakerParametersStructOutput,
    OptionGreekCache.GreekCacheParametersStructOutput,
    OptionGreekCache.ForceCloseParametersStructOutput,
    OptionGreekCache.MinCollateralParametersStructOutput,
    OptionMarketPricer.PricingParametersStructOutput,
    OptionMarketPricer.TradeLimitParametersStructOutput,
    OptionMarketPricer.VarianceFeeParametersStructOutput,
    OptionToken.PartialCollateralParametersStructOutput
  ] & {
    optionMarketParams: OptionMarket.OptionMarketParametersStructOutput;
    lpParams: LiquidityPool.LiquidityPoolParametersStructOutput;
    cbParams: LiquidityPool.CircuitBreakerParametersStructOutput;
    greekCacheParams: OptionGreekCache.GreekCacheParametersStructOutput;
    forceCloseParams: OptionGreekCache.ForceCloseParametersStructOutput;
    minCollatParams: OptionGreekCache.MinCollateralParametersStructOutput;
    pricingParams: OptionMarketPricer.PricingParametersStructOutput;
    tradeLimitParams: OptionMarketPricer.TradeLimitParametersStructOutput;
    varianceFeeParams: OptionMarketPricer.VarianceFeeParametersStructOutput;
    partialCollatParams: OptionToken.PartialCollateralParametersStructOutput;
  };
// ...
  export type MarketViewWithBoardsStructOutput = [
    boolean,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    LiquidityPool.LiquidityStructOutput,
    OptionMarketViewer.OptionMarketAddressesStructOutput,
    OptionMarketViewer.MarketParametersStructOutput,
    OptionGreekCache.NetGreeksStructOutput,
    OptionMarketViewer.BoardViewStructOutput[]
  ] & {
    isPaused: boolean;
    spotPrice: BigNumber;
    minSpotPrice: BigNumber;
    maxSpotPrice: BigNumber;
    quoteSymbol: string;
    quoteDecimals: BigNumber;
    baseSymbol: string;
    baseDecimals: BigNumber;
    rateAndCarry: BigNumber;
    liquidity: LiquidityPool.LiquidityStructOutput;
    marketAddresses: OptionMarketViewer.OptionMarketAddressesStructOutput;
    marketParameters: OptionMarketViewer.MarketParametersStructOutput;
    globalNetGreeks: OptionGreekCache.NetGreeksStructOutput;
    liveBoards: OptionMarketViewer.BoardViewStructOutput[];
  };
// ...
}

export declare namespace OptionGreekCache {
  export type NetGreeksStruct = {
    netDelta: PromiseOrValue<BigNumberish>;
    netStdVega: PromiseOrValue<BigNumberish>;
    netOptionValue: PromiseOrValue<BigNumberish>;
  };

  export type NetGreeksStructOutput = [BigNumber, BigNumber, BigNumber] & {
    netDelta: BigNumber;
    netStdVega: BigNumber;
    netOptionValue: BigNumber;
  };

  export type StrikeGreeksStruct = {
    callDelta: PromiseOrValue<BigNumberish>;
    putDelta: PromiseOrValue<BigNumberish>;
    stdVega: PromiseOrValue<BigNumberish>;
    callPrice: PromiseOrValue<BigNumberish>;
    putPrice: PromiseOrValue<BigNumberish>;
  };

  export type StrikeGreeksStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    callDelta: BigNumber;
    putDelta: BigNumber;
    stdVega: BigNumber;
    callPrice: BigNumber;
    putPrice: BigNumber;
  };

  export type GreekCacheParametersStruct = {
    maxStrikesPerBoard: PromiseOrValue<BigNumberish>;
    acceptableSpotPricePercentMove: PromiseOrValue<BigNumberish>;
    staleUpdateDuration: PromiseOrValue<BigNumberish>;
    varianceIvGWAVPeriod: PromiseOrValue<BigNumberish>;
    varianceSkewGWAVPeriod: PromiseOrValue<BigNumberish>;
    optionValueIvGWAVPeriod: PromiseOrValue<BigNumberish>;
    optionValueSkewGWAVPeriod: PromiseOrValue<BigNumberish>;
    gwavSkewFloor: PromiseOrValue<BigNumberish>;
    gwavSkewCap: PromiseOrValue<BigNumberish>;
  };

  export type GreekCacheParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    maxStrikesPerBoard: BigNumber;
    acceptableSpotPricePercentMove: BigNumber;
    staleUpdateDuration: BigNumber;
    varianceIvGWAVPeriod: BigNumber;
    varianceSkewGWAVPeriod: BigNumber;
    optionValueIvGWAVPeriod: BigNumber;
    optionValueSkewGWAVPeriod: BigNumber;
    gwavSkewFloor: BigNumber;
    gwavSkewCap: BigNumber;
  };

  export type ForceCloseParametersStruct = {
    ivGWAVPeriod: PromiseOrValue<BigNumberish>;
    skewGWAVPeriod: PromiseOrValue<BigNumberish>;
    shortVolShock: PromiseOrValue<BigNumberish>;
    shortPostCutoffVolShock: PromiseOrValue<BigNumberish>;
    longVolShock: PromiseOrValue<BigNumberish>;
    longPostCutoffVolShock: PromiseOrValue<BigNumberish>;
    liquidateVolShock: PromiseOrValue<BigNumberish>;
    liquidatePostCutoffVolShock: PromiseOrValue<BigNumberish>;
    shortSpotMin: PromiseOrValue<BigNumberish>;
    liquidateSpotMin: PromiseOrValue<BigNumberish>;
  };

  export type ForceCloseParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    ivGWAVPeriod: BigNumber;
    skewGWAVPeriod: BigNumber;
    shortVolShock: BigNumber;
    shortPostCutoffVolShock: BigNumber;
    longVolShock: BigNumber;
    longPostCutoffVolShock: BigNumber;
    liquidateVolShock: BigNumber;
    liquidatePostCutoffVolShock: BigNumber;
    shortSpotMin: BigNumber;
    liquidateSpotMin: BigNumber;
  };

  export type MinCollateralParametersStruct = {
    minStaticQuoteCollateral: PromiseOrValue<BigNumberish>;
    minStaticBaseCollateral: PromiseOrValue<BigNumberish>;
    shockVolA: PromiseOrValue<BigNumberish>;
    shockVolPointA: PromiseOrValue<BigNumberish>;
    shockVolB: PromiseOrValue<BigNumberish>;
    shockVolPointB: PromiseOrValue<BigNumberish>;
    callSpotPriceShock: PromiseOrValue<BigNumberish>;
    putSpotPriceShock: PromiseOrValue<BigNumberish>;
  };

  export type MinCollateralParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    minStaticQuoteCollateral: BigNumber;
    minStaticBaseCollateral: BigNumber;
    shockVolA: BigNumber;
    shockVolPointA: BigNumber;
    shockVolB: BigNumber;
    shockVolPointB: BigNumber;
    callSpotPriceShock: BigNumber;
    putSpotPriceShock: BigNumber;
  };
}

export declare namespace LiquidityPool {
  export type LiquidityStruct = {
    freeLiquidity: PromiseOrValue<BigNumberish>;
    burnableLiquidity: PromiseOrValue<BigNumberish>;
    reservedCollatLiquidity: PromiseOrValue<BigNumberish>;
    pendingDeltaLiquidity: PromiseOrValue<BigNumberish>;
    usedDeltaLiquidity: PromiseOrValue<BigNumberish>;
    NAV: PromiseOrValue<BigNumberish>;
    longScaleFactor: PromiseOrValue<BigNumberish>;
  };

  export type LiquidityStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    freeLiquidity: BigNumber;
    burnableLiquidity: BigNumber;
    reservedCollatLiquidity: BigNumber;
    pendingDeltaLiquidity: BigNumber;
    usedDeltaLiquidity: BigNumber;
    NAV: BigNumber;
    longScaleFactor: BigNumber;
  };

  export type LiquidityPoolParametersStruct = {
    minDepositWithdraw: PromiseOrValue<BigNumberish>;
    depositDelay: PromiseOrValue<BigNumberish>;
    withdrawalDelay: PromiseOrValue<BigNumberish>;
    withdrawalFee: PromiseOrValue<BigNumberish>;
    guardianMultisig: PromiseOrValue<string>;
    guardianDelay: PromiseOrValue<BigNumberish>;
    adjustmentNetScalingFactor: PromiseOrValue<BigNumberish>;
    callCollatScalingFactor: PromiseOrValue<BigNumberish>;
    putCollatScalingFactor: PromiseOrValue<BigNumberish>;
  };

  export type LiquidityPoolParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    string,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    minDepositWithdraw: BigNumber;
    depositDelay: BigNumber;
    withdrawalDelay: BigNumber;
    withdrawalFee: BigNumber;
    guardianMultisig: string;
    guardianDelay: BigNumber;
    adjustmentNetScalingFactor: BigNumber;
    callCollatScalingFactor: BigNumber;
    putCollatScalingFactor: BigNumber;
  };

  export type CircuitBreakerParametersStruct = {
    liquidityCBThreshold: PromiseOrValue<BigNumberish>;
    liquidityCBTimeout: PromiseOrValue<BigNumberish>;
    ivVarianceCBThreshold: PromiseOrValue<BigNumberish>;
    skewVarianceCBThreshold: PromiseOrValue<BigNumberish>;
    ivVarianceCBTimeout: PromiseOrValue<BigNumberish>;
    skewVarianceCBTimeout: PromiseOrValue<BigNumberish>;
    boardSettlementCBTimeout: PromiseOrValue<BigNumberish>;
    contractAdjustmentCBTimeout: PromiseOrValue<BigNumberish>;
  };

  export type CircuitBreakerParametersStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber
  ] & {
    liquidityCBThreshold: BigNumber;
    liquidityCBTimeout: BigNumber;
    ivVarianceCBThreshold: BigNumber;
    skewVarianceCBThreshold: BigNumber;
    ivVarianceCBTimeout: BigNumber;
    skewVarianceCBTimeout: BigNumber;
    boardSettlementCBTimeout: BigNumber;
    contractAdjustmentCBTimeout: BigNumber;
  };
}